JIT: Handle possibility of late optimized out async calls
If we optimize an async call away after the suspension/resumption code
has been created then the emit locations used for resumption info and
diagnostic info will not be valid. Handle this rare case by just storing
0. The value should not matter as we will never suspend here.

Ideally we would model things so that suspension/resumption blocks could
too be removed in these cases, but that's not so simple (we do not have
a mechanism to represent a data dependency of a basic block, only a
control flow dependency).